Files that can't be parsed are no longer silently skipped
It failed some of the specs and gave errors for the wrong
reasons: the content couldn't be parsed as Ruby rather than the check
under test actually failing for the right reason.

If I ran Roodi against a file that wasn't valid Ruby, I'd like to know,
rather than getting a "Found 0 errors" message.